requirements.txt
psycopg2-binary
aiogram
python-dotenv
